\name{Genera3}
\alias{Genera3}
\docType{data}
\title{
Taxa count data set of genera Bacteroids, Prevetella and Ruminococcus.
}
\description{
In the cross-sectional study of diet and stool microbiome composition (Wu et al., 2011), stool samples from 98 healthy volunteers were collected. the DNA samples were analyzed and 3,608 OTUs are defined. These OTUs were further combined into 11 phyla and 127 genera. The Genera3 data set contains taxa counts of the enterotype genera Bacteroids, Prevetella and Ruminococcus for the 98 individuals.
}
\usage{data(Genera3)}
\format{
   The format is:
 chr "Genera3"
}

\source{
1. Gary D Wu, Jun Chen, Christian Hoffmann, Kyle Bittinger, Ying-Yu Chen, Sue A. Keilbaugh, et al (2011). Linking long-term dietary patterns with gut microbial enterotypes. Science.

2. Jun Chen and Hongzhe Li (2012). Variable selection for sparse Dirichlet-Multinomial regression with an application to microbiome data analysis. Annals of Applied Statistics, (in press).

3. Fan Xia, Jun Chen, Wing Kam Fung, Hongzhe, Li (2012). A Logistic Normal Multinomial Regression Model for Microbiome Compositional Data Analysis. (Submitted)
}

\examples{
data(Genera3)
}
\keyword{datasets}
