\name{Nutrients}
\alias{Nutrients}
\docType{data}
\title{
Data matrix of nutrients
}
\description{
Information on habitual long-term diet of 98 subjects was collected using the food frequency questionnaire (FFQ), which results in quantitative measurements of 214 micro-nutrients. These micro-nutrients are further grouped together if the correlation coefficient > 0.9. For each covariate group, one representative nutrient is selected, leading to a total of p = 117 covariates in the \strong{Nutrients} data set.
}
\usage{data(Nutrients)}
\format{
 The format is:
 chr "Nutrients"
}

\source{
1. Gary D. Wu, Jun Chen, Christian Hoffmann, Kyle Bittinger, Ying-Yu Chen, Sue A. Keilbaugh, et al (2011). Linking long-term dietary patterns with gut microbial enterotypes. Science.

2. Jun Chen and Hongzhe Li (2012). Variable selection for sparse Dirichlet-Multinomial regression with an application to microbiome data analysis. Annals of Applied Statistics, (in press).

3. Fan Xia, Jun Chen, Wing Kam Fung, Hongzhe, Li (2012). A Logistic Normal Multinomial Regression Model for Microbiome Compositional Data Analysis. (Submitted)
}

\examples{
data(Nutrients)
}
\keyword{datasets}
